minor changes to structure and Ecdsa doc test
