+++ Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "universal_wallet"
-version = "0.3.2"
+version = "0.3.3"
 authors = ["Charles Cunningham <c.a.cunningham6@gmail.com>",
             "Ivan Temchenko <35359595i@gmail.com>"]
 edition = "2018"
